Do not take 100% of the CPU in gksu-polkit when there's no stdin
If there is not stdin (or if it gets closed), we get a EOF, and we
should just stop trying to receive input, instead of trying again and
again.

Also, do not queue anything in the pipe to the stdin of the child if
what we got through our stdin is empty.